2024-04-24 install script - master branch
The installer script checks for `/boot/cmdline.txt`, assumes its
presence implies the Raspberry Pi and proceeds to add options to it (to
support `docker stats`). With Bookworm, `/boot/cmdline.txt` is a read-me
and the actual boot options file is at `/boot/firmware/cmdline.txt`.

This mod checks for the latter first, then falls back to checking for
the former.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>